Sphere:nbRefPoints is a variable, not a method Warnings use \n Added explicit casts to silence warnings

+++ src/base/ansi_colors.cc
@@ -35,7 +35,7 @@
 /* Check the number of colors that the terminal supports */
 bool has_colors()
 {
-    static int n_colors = 0;
+    static long n_colors = 0;
     if ( n_colors == 0 )
     {
         long n = 1;
